qc 적용

+++ teamProject/teamProject/teamProject/QData.cs
@@ -9,10 +9,10 @@ namespace teamProject
     public class QData
     {
         public DateTime date { get; set; } // PK
-        public int weight { get; set; }
+        public double weight { get; set; }
         public double water { get; set; }
         public double material { get; set; }
-        public int HSO { get; set; }
+        public double HSO { get; set; }
         public double pH { get; set; }
     }
 }
